Import upstream ceph mgr module
Since the upstream doesn't provide a pyproject.toml or setup.py
it causes problems when trying to use as a dependency. Instead,
we're going to to try and just import the upstream sources.

This was done using the git diff | git apply method outlined in
https://stackoverflow.com/questions/23937436/add-subdirectory-of-remote-repo-with-git-subtree
with the notable exception that I did not keep the git history
in order to avoid the git history polluting the history of this
repository. That may later prove to be a mistake - but one that
can be rectified later.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>